Fix for suspensions started in era >= 1000 (#142)
Fix for azero.dev bug, which occurs for eras >= 1000. Is caused by
invalid to number cast. `toHuman` casts numbers with more than 4 digits
using comma thousands separator, e.g. `1008` era is cast as `1,008`,
which then `toNumber` was not able to handle. It was clear that
suspensions was not correctly parsed, ie without explicit types. Now it
should be better. See backend structs:
https://github.com/Cardinal-Cryptography/aleph-node/blob/release-14/primitives/src/lib.rs#L246-L262

This change is dedicated for Mainnet and Testnet release, so it is based
on `release-11` branch.